requirements.txt
streamlit==1.31.0
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
moviepy==1.0.3
langchain==0.1.0 
